refactor(semantics): place action helpers beside their callers
Follow the function-placement rule: a function lives next to its use,
with visibility equal to its use.

- missingText/missingIndex/performed/notSupported become factories on
  the result type in NodeActionResults.kt (missingArgument,
  performedIf, notSupported); messages are unchanged.
- SemanticsConfiguration.invokeAction leaves the interface file for
  SemanticsConfigurationActions.kt.
- The BringIntoView walk and its private helpers move into
  ScrollActions.BringIntoView; scrollDeltaToReveal stays internal on
  the object for its test.
- hasItemPositions and isRecyclerViewAvailable move into
  ViewScrollActions.ScrollToIndex, their only user.
- Rect.toOutwardAndroidRect moves next to requestRectangleOnScreenNow
  in AndroidWindowNodeSource.kt as a private function.
