+++ jetwhale-plugins/semantics/agent/src/androidMain/kotlin/com/kitakkun/jetwhale/plugins/semantics/agent/ViewScrollActions.kt
@@ -4,6 +4,7 @@ import android.graphics.Rect
 import android.view.View
 import android.widget.AbsListView
 import androidx.recyclerview.widget.RecyclerView
+import com.kitakkun.jetwhale.plugins.semantics.protocol.NodeAction
 import com.kitakkun.jetwhale.plugins.semantics.protocol.NodeActionResult
 import com.kitakkun.jetwhale.plugins.semantics.protocol.PerformNodeAction
 import kotlin.math.roundToInt
@@ -16,7 +17,7 @@ internal object ViewScrollActions {
         override fun isOfferedBy(view: View) = view.isScrollable()
 
         override fun perform(view: View, request: PerformNodeAction): NodeActionResult {
-            if (!view.isScrollable()) return notSupported("the view has nothing to scroll")
+            if (!view.isScrollable()) return NodeActionResult.notSupported("the view has nothing to scroll")
             view.scrollBy(request.scrollX.roundToInt(), request.scrollY.roundToInt())
             return NodeActionResult(performed = true)
         }
@@ -31,14 +32,14 @@ internal object ViewScrollActions {
     object ScrollToIndex : ViewActionHandler {
         override val runsOnDisabledView = true
 
-        override fun isOfferedBy(view: View) = view.hasItemPositions()
+        override fun isOfferedBy(view: View) = view is AbsListView || (isRecyclerViewAvailable && view is RecyclerView)
 
         override fun perform(view: View, request: PerformNodeAction): NodeActionResult {
-            val index = request.index ?: return missingIndex()
+            val index = request.index ?: return NodeActionResult.missingArgument(NodeAction.ScrollToIndex, "index")
             return when {
                 view is AbsListView -> {
                     if (index !in 0 until view.count) {
-                        notSupported("index $index is out of bounds [0, ${view.count})")
+                        NodeActionResult.notSupported("index $index is out of bounds [0, ${view.count})")
                     } else {
                         view.setSelection(index)
                         NodeActionResult(performed = true)
@@ -48,15 +49,23 @@ internal object ViewScrollActions {
                 isRecyclerViewAvailable && view is RecyclerView -> {
                     val itemCount = view.adapter?.itemCount ?: 0
                     if (index !in 0 until itemCount) {
-                        notSupported("index $index is out of bounds [0, $itemCount)")
+                        NodeActionResult.notSupported("index $index is out of bounds [0, $itemCount)")
                     } else {
                         view.scrollToPosition(index)
                         NodeActionResult(performed = true)
                     }
                 }
 
-                else -> notSupported("the view is not a RecyclerView or a ListView")
+                else -> NodeActionResult.notSupported("the view is not a RecyclerView or a ListView")
             }
+        }
+
+        /**
+         * An `is RecyclerView` check on an app that does not ship the library fails to resolve the
+         * class and throws instead of answering `false`, so the class is looked up by name first.
+         */
+        private val isRecyclerViewAvailable: Boolean by lazy {
+            runCatching { Class.forName("androidx.recyclerview.widget.RecyclerView") }.isSuccess
         }
     }
 
@@ -73,12 +82,12 @@ internal object ViewScrollActions {
 
         override fun perform(view: View, request: PerformNodeAction): NodeActionResult {
             val whole = Rect(0, 0, view.width, view.height)
-            if (whole.isEmpty) return notSupported("the view has no size to bring into view")
+            if (whole.isEmpty) return NodeActionResult.notSupported("the view has no size to bring into view")
             val visible = Rect()
             if (view.getLocalVisibleRect(visible) && visible == whole) {
                 return NodeActionResult(performed = true, message = "the view is already in view")
             }
-            return performed(view.requestRectangleOnScreen(whole, true), "no ancestor scrolled the view into view")
+            return NodeActionResult.performedIf(view.requestRectangleOnScreen(whole, true), "no ancestor scrolled the view into view")
         }
     }
 }
