+++ jetwhale-plugins/semantics/agent/src/androidMain/kotlin/com/kitakkun/jetwhale/plugins/semantics/agent/ViewTreeCapture.kt
@@ -4,7 +4,6 @@ import android.content.res.Resources
 import android.graphics.Rect
 import android.view.View
 import android.view.ViewGroup
-import android.widget.AbsListView
 import android.widget.Checkable
 import android.widget.EditText
 import android.widget.TextView
@@ -13,15 +12,12 @@ import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.node.InteroperableComposeUiNode
 import androidx.compose.ui.platform.ViewRootForTest
 import androidx.compose.ui.semantics.SemanticsNode
-import androidx.recyclerview.widget.RecyclerView
 import com.kitakkun.jetwhale.plugins.semantics.protocol.NodeAction
 import com.kitakkun.jetwhale.plugins.semantics.protocol.NodeBounds
 import com.kitakkun.jetwhale.plugins.semantics.protocol.NodeTreeCaptureOptions
 import com.kitakkun.jetwhale.plugins.semantics.protocol.UiNode
 import com.kitakkun.jetwhale.plugins.semantics.protocol.ViewNode
 import com.kitakkun.jetwhale.plugins.semantics.protocol.advertisedAs
-import kotlin.math.ceil
-import kotlin.math.floor
 
 /**
  * Converts an Android `View` subtree into the transport model, descending into every composition it
@@ -146,18 +142,6 @@ private fun View.viewActionNames(): List<String> = NodeAction.entries.mapNotNull
     action.advertisedAs?.takeIf { action.viewHandler.isOfferedBy(this) }
 }
 
-/** Whether the view is a list that can be told which item to show — the View side of `ScrollToIndex`. */
-internal fun View.hasItemPositions(): Boolean = this is AbsListView || (isRecyclerViewAvailable && this is RecyclerView)
-
-/**
- * The agent compiles against `RecyclerView` without bundling it, so an app that does not ship the
- * library must never reach an `is RecyclerView` check: the class fails to resolve and the check
- * throws instead of answering `false`.
- */
-internal val isRecyclerViewAvailable: Boolean by lazy {
-    runCatching { Class.forName("androidx.recyclerview.widget.RecyclerView") }.isSuccess
-}
-
 /**
  * A view scrolls when it has somewhere to scroll to in any direction. Asking the view itself covers
  * every scrolling container — `ScrollView`, `RecyclerView`, a custom one — without naming any.
@@ -217,15 +201,4 @@ private fun NodeBounds.translated(offsetX: Float, offsetY: Float): NodeBounds = 
     top = top + offsetY,
     right = right + offsetX,
     bottom = bottom + offsetY,
-)
-
-/**
- * The smallest integer rectangle that covers [this] — a request to show it must not lose the
- * fraction of a pixel at either edge.
- */
-internal fun androidx.compose.ui.geometry.Rect.toOutwardAndroidRect(): Rect = Rect(
-    floor(left).toInt(),
-    floor(top).toInt(),
-    ceil(right).toInt(),
-    ceil(bottom).toInt(),
 )
